+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 - @cosmjs/launchpad: `Secp256k1HdWallet.sign` method renamed `signAmino`.
 - @cosmjs/launchpad: Add `CosmosClient.getTx` method for searching by ID and
   remove such functionality from `CosmosClient.searchTx`.
+- @cosmjs/launchpad: Add `SigningCosmosClient.sign` method for signing without
+  broadcasting.
 - @cosmjs/launchpad-ledger: `LedgerSigner.sign` method renamed `signAmino`.
 - @cosmjs/proto-signing: Add new package for handling transaction signing with
   protobuf encoding.